refactor(rw): split rw into dedicated submodule files
