+++ src/core/rw/libxr_rw.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "operation.hpp"
+#include "read_port.hpp"
+#include "write_port.hpp"
+#include "stdio.hpp"
